Task details page: clean up common elements, add specific code for TaskDetails only
