+++ src/client/modules/Tasks/TaskDetails/index.jsx
@@ -2,12 +2,21 @@ import React from 'react'
 import { useParams } from 'react-router-dom'
 import { connect } from 'react-redux'
 
-import { DefaultLayout } from '../../../components'
+import styled from 'styled-components'
+
+import { SPACING } from '@govuk-react/constants'
+
+import { DefaultLayout, Tag } from '../../../components'
 import { buildCompanyBreadcrumbs } from '../../Companies/utils'
 import TaskDetailsTable from './TaskDetailsTable'
 import { ID, TASK_GET_TASK_DETAILS, state2props } from './state'
 import { TASK_DETAILS_LOADED } from '../../../actions'
 import Task from '../../../components/Task'
+
+const StyledTag = styled(Tag)`
+  display: 'block',
+  margin-bottom: ${SPACING.SCALE_2};
+`
 
 const getCompany = (task) => {
   return task?.investmentProject?.investorCompany
@@ -26,8 +35,18 @@ const TaskDetails = ({ task }) => {
 
   return (
     <DefaultLayout
-      heading={taskTitle}
-      archived={archived}
+      heading={{
+        ...(archived ? (
+          <div>
+            <h1>{taskTitle}</h1>
+            <StyledTag colour="green" data-test="activity-kind-label">
+              COMPLETED
+            </StyledTag>
+          </div>
+        ) : (
+          <h1>{taskTitle}</h1>
+        )),
+      }}
       pageTitle={taskTitle}
       breadcrumbs={buildCompanyBreadcrumbs(
         {
